+++ Library/src/Systems/Assets/AssetSystem.cpp
@@ -848,7 +848,7 @@ void AssetSystem::GetAssetsByCollectionIds(const Array<String>& AssetCollectionI
 
     std::vector<String> Ids;
 
-    for (int i = 0; i < AssetCollectionIds.Size(); ++i)
+    for (size_t i = 0; i < AssetCollectionIds.Size(); ++i)
     {
         Ids.push_back(AssetCollectionIds[i]);
     }
@@ -1001,7 +1001,7 @@ CSP_ASYNC_RESULT_WITH_PROGRESS void AssetSystem::RegisterAssetToLODChain(
         Asset NewAsset = InAsset;
         Array<String> NewStyles(NewAsset.Styles.Size() + 1);
 
-        for (int i = 0; i < NewAsset.Styles.Size(); ++i)
+        for (size_t i = 0; i < NewAsset.Styles.Size(); ++i)
         {
             NewStyles[i] = NewAsset.Styles[i];
         }
